% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{alnSeq2seqMatrix}
\alias{alnSeq2seqMatrix}
\title{Convert an alignment from a vector of character strings to a matrix}
\usage{
alnSeq2seqMatrix(seqDT)
}
\arguments{
\item{seqDT}{A data.table, meeting the specs for a seqDT.  [no default]}
}
\value{
A numeric value
}
\description{
Useful for column-wise calculations (like consensus sequences, minor allele frequencies, etc.)
}
\details{
One row per sequence, one entry per alignment column (i.e. per nucleotide/amino, or gap). Row names are taken from the column 'seqId'. The method is sequence-agnostic; any characters trings of equal length will work.
}
\examples{
He(c("A","A","G"))
}
