% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{ask}
\alias{ask}
\title{Collect input from a user interactively.}
\usage{
ask(question = "Please enter a value", YN = FALSE)
}
\arguments{
\item{question}{[optional] A question to prompt the user with ["Please enter a value"]}

\item{YN}{[optional] Does the question require a Yes/No answer? If so, the question will be ended with "[Y/N]" and only the responses "Y" and "N" will be accepted--and will return logical `TRUE` or `FALSE` values accordingly. [NULL]}
}
\value{
Whatever the user inputs before pressing <return>, as a character string.
}
\description{
Collect input from a user interactively.
}
\examples{
while(ask("Guess the number I'm thinking of!")!=sample(1:1e9L,1)){ ce("WRONG!") }
}
