% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_align_blast.R
\name{blast}
\alias{blast}
\title{Call blast* aligner family programs}
\usage{
blast(
  subjectFname = NULL,
  queryFname = NULL,
  subjectFastaDT = NULL,
  queryFastaDT = NULL,
  outFastaFname = NULL,
  program = "blastn",
  blastFmtArgs = "",
  outFmtArg =
    "6       saccver  qaccver slen qlen length sstart send qstart qend pident evalue bitscore",
  outputColNames = c("sSeqId", "qSeqId", "sLength", "qLength", "matchLength", "sStart",
    "sEnd", "qStart", "qEnd", "pctId", "eValue", "bitscore"),
  outputColClasses = c("character", "character", "integer", "integer", "integer",
    "numeric", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric"),
  addAlignments = FALSE,
  moreBlastArgs = "",
  blastBinaryDir = sub("/blastn$", "", system("which blastn", intern = T)),
  makeBlastDbBinary = system("which makeblastdb", intern = T)
)
}
\arguments{
\item{param1_name}{What it is. Defaults?}

\item{param2_name}{What it is. Defaults?}
}
\value{
What the function returns.
}
\description{
Describe what the function does in detail.
}
\examples{
Example 1 code

Example 2 code
}
