% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{charVec2Matrix}
\alias{charVec2Matrix}
\title{#' @export
Convert a vector of equal-length character strings into a matrix.}
\usage{
charVec2Matrix(x)
}
\arguments{
\item{x}{A (most usefully, character) vector.}
}
\value{
A character matrix
}
\description{
One string per row, one char per entry.
}
\examples{
charVec2Matrix(c("Hello","dolly","sings","Satch"))
}
