% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{existentDir}
\alias{existentDir}
\title{Use it in place of a dirname to ensure it really exists.}
\usage{
existentDir(dirName)

existentDir(dirName)
}
\description{
Use it in place of a dirname to ensure it really exists.

Return the argument (a directory name) but also make sure a directory called that exists.
}
