% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_extractSeq.R
\name{extractSeq}
\alias{extractSeq}
\title{Extract the subsequences described by a coordinate table (typically a bed file, bed dt or similar), from a sequence list (e.g. fasta file or dt).}
\usage{
extractSeq(
  bedFname = NULL,
  bedDT = NULL,
  fastaFname = NULL,
  fastaDT = NULL,
  outFastaFname = NULL,
  stranded = TRUE,
  nameCol = NULL,
  bedToolsBin = system("which bedtools", intern = T)
)
}
\arguments{
\item{paramName}{What it is. Defaults?}
}
\value{
What the function returns.
}
\description{
Describe what the function does in detail.
}
\examples{
Example 1 code

Example 2 code
}
