% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_fai.R
\name{getFai}
\alias{getFai}
\title{Create a .fai-flavoured data.table directly from a .fasta file}
\usage{
getFai(fastaFname, samtoolsBin = system("which samtools", intern = TRUE))
}
\arguments{
\item{fastaFname}{A character vector. The .fasta filename(s). [no default]}

\item{samtoolsBin}{Character value. In case namesOnly==TRUE, where to find samtools. [system("which samtools", intern=TRUE)]}
}
\value{
A data.table with .fai flavour.
}
\description{
Calls samtools faidx
}
