% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{\%plusMinus\%}
\alias{\%plusMinus\%}
\title{Return the range of x, extended by a bit}
\usage{
x \%plusMinus\% m
}
\arguments{
\item{x}{The vector to take the range of}

\item{m}{Subtract me from the low end}

\item{p}{Add me to the high end [m]}
}
\value{
A length two vector
}
\description{
Return the range of x, extended by a bit
}
