% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{gridApplyDT}
\alias{gridApplyDT}
\title{Create a grid of values, calculated from every combination of two rows, one drawn from each of two data.tables.}
\usage{
gridApplyDT(dtx, dty = NULL, FUN, nameColx = NULL, nameColy = NULL)
}
\arguments{
\item{dtx}{A data table. [no default]}

\item{dty}{[optional] A data.table. If omitted, [NULL]}

\item{FUN}{[no default]}

\item{nameColx}{[optional] Name of the column in `dtx` from which the output row names are to be taken. [NULL]}

\item{nameColy}{[optional] Name of the column in `dty` from which the output column names are to be taken. [NULL]}
}
\value{
A matrix of class whatever FUN outputs.
}
\description{
The i,j-th entry of the output matrix is the result of `FUN(dtx[i],dty[j])`.
}
\details{
Wraps to the console width. Coordinates are given relative to the sequence start.
}
\examples{
ldply but a data.table returns
d1 <- d.t(value=1:26,label1=LETTERS,label2=letters)
d2 <- d.t(value=26:1,label1=LETTERS,label2=letters)
gridApplyDT(d1,nameColx = "label1",nameColy = "label2",FUN = function(x,y){ abs(x$value - y$value) })
gridApplyDT(d1,d2,nameColx = "label1",nameColy = "label2",FUN = function(x,y){ abs(x$value - y$value) })
}
