% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{He}
\alias{He}
\title{Effective heterozygosity of a vector of alleles}
\usage{
He(x, inclNAs = F)
}
\arguments{
\item{x}{A  vector. Canonically, listing alleles in a population at some site. [no default]}

\item{inclNAs}{A logical value. Delete NA entries before making He calculation? [FALSE]}
}
\value{
A numeric value
}
\description{
Calculates the probability that two randomly drawn alleles will be different.
}
\details{
Does not account for ploidy in any way. Is agnostic to what the given alleles are. c("A","G","C","T","-") will give the c("same","result","as","this","would"). No characters including gaps ("-") or NAs are recognised as special (by default).
}
\examples{
He(c("A","A","G"))
}
