% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{killGaps}
\alias{killGaps}
\title{Delete '-' characters from strings}
\usage{
killGaps(x)
}
\arguments{
\item{x}{A character vector.  [no default]}
}
\value{
A character vector the same length of x, with all '-' characters omitted.
}
\description{
Delete '-' characters from strings
}
\details{
A common use is to delete gaps from aligned sequences to they can become raw sequences again, and thus aligned to yet other sequences etc.
}
\examples{
killGaps("AGCT--No-More-Gaps--TCGA")
}
