% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_fai.R
\name{makeFai}
\alias{makeFai}
\title{Generate .fai files for given .fasta files.}
\usage{
makeFai(
  fastaFname,
  faiFname = paste0(fastaFname, ".fai"),
  samtoolsBin = system("which samtools", intern = TRUE)
)
}
\arguments{
\item{fastaFname}{A character vector. The input .fasta filename(s). [no default]}

\item{faiFname}{A character vector. The output .fai filename(s). [paste0(fastaFname,".fai")]}

\item{samtoolsBin}{A character vector. The .fai filename. [system("which samtools", intern=TRUE)]}
}
\value{
A data.table with .fai flavour.
}
\description{
Wrapper for samtools faidx
}
