% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_palette.R
\name{makePalette}
\alias{makePalette}
\title{Makes a simple linear palette interpolating the colChain across n evenly spaced points. Mostly useful for calling from other funs.}
\usage{
makePalette(
  colChain = palettePresets$wheel,
  n = NULL,
  at = NULL,
  setAlpha = NULL
)
}
\description{
Makes a simple linear palette interpolating the colChain across n evenly spaced points. Mostly useful for calling from other funs.
}
