% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{plotAlignmentCartesian}
\alias{plotAlignmentCartesian}
\title{Cartesian alignment plots}
\usage{
plotAlignmentCartesian(
  alignmentDT,
  newPlot = TRUE,
  bgColour = "#000011",
  colourAlignments = NULL,
  alignmentColourDefault = "#FFFFEE",
  lwd = 1,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{alignmentDT}{A data.table, meeting the specs for an alignmentDT. [no default]}

\item{newPlot}{Add to an existing plot, or create a new one? [FALSE]}

\item{bgColour}{["#000011", aka a very dark navy blue]}

\item{colourAlignments}{If the alignments are to be coloured, then name the column from which to draw them. See examples. [NULL]}

\item{alignmentColourDefault}{If 'colourAlignments' isn't used, then what colour should they be? ["#FFFFEE", aka light cream]}

\item{lwd}{The line width [1.0]}

\item{xlim}{Use to limit the range along x (the 'subject sequence' axis) [NULL]}

\item{ylim}{Use to limit the range along x (the 'subject sequence' axis) [NULL]}
}
\value{
Nothing, it draws a plot.
}
\description{
Often called a dotplot for historical (read "invalid") reasons. That's a stupid name since the word dot just doesn't relate to anything going on here in any way. Hence, the shiny new name "cartesian alignment plot" (c.f., say, Sankey alignment plots).
}
\details{
This assumes you've included just one subject and query seq (i.e. `nu(c(sSeqId,qSeqId))==2`). In future this check may be made explicit. If you do include more, the alignments will all be overlapped on one plot, which is not likely to be what you intended.
}
\examples{
Coming ...
}
