% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{plotSequenceAlignment}
\alias{plotSequenceAlignment}
\title{Plots alignments with coloured nucleotides}
\usage{
plotSequenceAlignment(
  seqDT,
  rowGap = 0.05,
  alnPreset = c("guess", "dna", "aa"),
  alnItems = NULL,
  alnItemCols = NULL,
  invariantColour = NULL,
  gapColour = "#BBBBBB",
  unknownNtColour = "#FF55FF",
  labelsInAxis = TRUE,
  cex.axis = 0.7,
  ...
)
}
\arguments{
\item{seqDT}{A data.table, meeting the specs for a seqDT. [no default]}

\item{rowGap}{The size of the gap between alignment rows [0.05]}
}
\value{
A numeric value
}
\description{
Alignments are plotted horizontally, with the first given sequence on top.
}
\details{
To plot from an alignmentDT, do a manual conversion. E.g., to 'melt' the alignments:
```alignmentDT[,alnId:=paste0("Alignment_",1:.N)][,seqId=c(sSeqId,qSeqId),alnSeq=c(sAlnSeq,qAlnSeq),by=.(alnId)]```
or call `alignedSeqDT2seqMatrix()` (be aware this function preserves only one of the subject or query sequences). Also, be aware most alignmentDTs have different length alignments, so you will typically be selecting a limited number of rows.
}
\examples{
seqDT <- data.table(
  seq = c("GC-AT",
          "TC-AA"),
  seq2 = c("ABCD","EFGH"),
  seqId = LETTERS[1:2]
)
plotNucleotideAlignment(seqDT)
plotNucleotideAlignment(seqDT,invariantColour = "black")
plotNucleotideAlignment(seqDT,alnItemCols = applyPalette(1:4,palettePresets$williams),gapColour = "white")
plotNucleotideAlignment(seqDT[,.(seqId,seq=seq2)])
}
