% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{printAln}
\alias{printAln}
\title{Print sequences as an alignment}
\usage{
printAln(s)
}
\arguments{
\item{x}{A character vector. With equal-length aligned DNA sequences if the results are to make sense. [no default]}
}
\description{
A BLAST-esque text alignment with bars ('|') showing matches between neighbouring sequences.
}
\details{
Wraps to the console width. Coordinates are given relative to the sequence starts.
}
\examples{
printAln(c(
  "AAAGGCTGAGTCGAGTC",
  "AAAxxCT--GxCGxG-C"
))
}
