% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_bioinfo.R
\name{rc}
\alias{rc}
\title{Reverse Complement}
\usage{
rc(x)
}
\arguments{
\item{x}{A character vector. With DNA sequences if the results are to make sense. [no default]}
}
\value{
A character vector.
}
\description{
Calculate the reverse complements of DNA sequences.
}
\details{
Accepts (and preserves) uppercase and lowercase nucleotide codes, including IUPAC codes. Other characters will be left as-is.
}
\examples{
rc("acgt---ACGT---RYSWKMBDHVN---ryswkmbdhvn")
rc(rc(c("AaA-CcC","GgG-TtT")))
}
