% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_bed.R
\name{readBed}
\alias{readBed}
\title{Read a bed file}
\usage{
readBed(bedFname, saveFnames = TRUE)
}
\arguments{
\item{bedFname}{A character vector. The .bed filenames to read. Calls awk. [no default]}

\item{saveFnames}{Logical value. Include a bedFname column in the output? [TRUE]}
}
\value{
A data.table with .bed flavour.
}
\description{
Read a bed file
}
