% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_gff.R
\name{readGff}
\alias{readGff}
\title{Read annotation files in gff format into a data.table}
\usage{
readGff(gffFname)
}
\arguments{
\item{faiFname}{A character vector. The .gff filename(s). [no default]}
}
\value{
A data.table with coordDT flavour.
}
\description{
Read annotation files in gff format into a data.table
}
