% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_alignToVarTable.R
\name{varTableFromAlns}
\alias{varTableFromAlns}
\title{Convert homology search/alignment output to a variant table DT.}
\usage{
varTableFromAlns(aln)
}
\arguments{
\item{aln}{What it is. Defaults?}
}
\value{
What the function returns.
}
\description{
Describe what the function does in detail.
}
\examples{
Example 1 code

Example 2 code
#' @export
}
