% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_bed.R
\name{writeBed}
\alias{writeBed}
\title{Write a bed file}
\usage{
writeBed(coordDT, bedFname = NULL)
}
\arguments{
\item{bedFname}{A character vector. The filename to write. If NULL will use the deparsed/substituted variable name given as bed [NULL]}

\item{bed}{A .fasta-flavoured data.table.}
}
\description{
Write a bed file
}
