% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_fasta.R
\name{writeFasta}
\alias{writeFasta}
\title{Write a fasta or fasta.gz file}
\usage{
writeFasta(seqDT, outFastaFname = NULL, gzip = F, nameCol = "seqId")
}
\arguments{
\item{gzip}{Logical value. Create gzipped outpit? Calls gzip via R's native gzfile() interface. [FALSE]}

\item{nameCol}{Character value. The column to take fasta header names from. ["seqId"]}

\item{fasta}{A .fasta-flavoured data.table.}

\item{fastaFname}{A character vector. The .fasta filenames to write. If NULL will use the deparsed/substituted variable name given as the arg to `fasta` [NULL]}
}
\description{
Calls awk. Note this writes all the given fasta-dts into ONE file, i.e., if you writeFasta(readFasta(someFastaFileNames)) you will concatenate all the someFastaFileNames into one output.
}
