% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{ask}
\alias{ask}
\title{Collect input from a user interactively.}
\usage{
ask(question = "NULL")
}
\arguments{
\item{question}{[optional] A question to prompt the user with [NULL]}
}
\value{
Whatever the user inputs before pressing <return>, as a character string.
}
\description{
Collect input from a user interactively.
}
\examples{
while(ask("Guess the number I'm thinking of!")!=sample(1:1e9L,1)){ ce("WRONG!") }
}
