% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_blast.R
\name{blast}
\alias{blast}
\title{Call blast* aligner family programs}
\usage{
blast(
  subjectFname = NULL,
  queryFname = NULL,
  subjectSeqDT = NULL,
  querySeqDT = NULL,
  outFastaFname = NULL,
  program = "blastn",
  addAlignments = FALSE,
  outFmtArg =
    "6       saccver     qaccver       length          sstart      send       qstart      qend       slen         qlen        pident           evalue     bitscore",
  outputColNames = c("sSeqId", "qSeqId", "matchLength", "sStart", "sEnd", "qStart",
    "qEnd", "sLength", "qLength", "pctId_noGaps", "eValue", "bitscore"),
  outputColClasses = c("character", "character", "integer", "numeric", "numeric",
    "numeric", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric"),
  moreBlastArgs = "",
  blastBinaryDir = sub("/blastn$", "", system("which blastn", intern = T)),
  makeBlastDbBinary = system("which makeblastdb", intern = T)
)
}
\arguments{
\item{param1_name}{What it is. Defaults?}

\item{param2_name}{What it is. Defaults?}
}
\value{
What the function returns.
}
\description{
Describe what the function does in detail.
}
\examples{
Example 1 code

Example 2 code
}
