% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_misc.R
\name{plusAtinyBit}
\alias{plusAtinyBit}
\title{Add a tiny bit to each number.}
\usage{
plusAtinyBit(x, inc = min(x) * 0.1)
}
\arguments{
\item{x}{The vector to add a bit to [no default]}

\item{inc}{How much to add [min(x)*0.1]}
}
\value{
A vector
}
\description{
Useful when a collection of numbers on [0,Inf) need to be logged
}
