% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_fasta.R
\name{readFasta}
\alias{readFasta}
\title{Read a fasta or fasta.gz file}
\usage{
readFasta(
  fastaFname,
  saveFnames = TRUE,
  namesOnly = FALSE,
  awkBin = system("which awk", intern = TRUE),
  samtoolsBin = system("which samtools", intern = TRUE)
)
}
\arguments{
\item{fastaFname}{A character vector. The .fasta filenames to read. Calls awk. [no default]}

\item{saveFnames}{Logical value. Include a fastaFname column in the output? [TRUE]}

\item{namesOnly}{Logical value. If true, will return only the seqName column. Calls samtools faidx.}

\item{awkBin}{Character value. Where to find awk [system("which awk", intern=TRUE)]}

\item{samtoolsBin}{Character value. In case namesOnly==TRUE, where to find samtools. [system("which samtools", intern=TRUE)]}

\item{gzipBin}{Character value. Where to find gzip. [system("which gzip", intern=TRUE)]}
}
\value{
A data.table with .fasta flavour.
}
\description{
Calls awk normally, and samtools faidx is namesOnly==TRUE. Calls gzip if the filename ends in ".gz".
}
