% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_gff.R
\name{readGff}
\alias{readGff}
\title{Read .fai files into a data.table}
\usage{
readGff(gffFname)
}
\arguments{
\item{faiFname}{A character vector. The .fai filename(s). [no default]}
}
\value{
A data.table with .fai flavour.
}
\description{
Read .fai files into a data.table
}
